+++ CHANGELOG.md
@@ -12,3 +12,4 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - Prevent writing when the previous time is stale, i.e. already edited in another browser window
+- Don't re-render the entire history page when editing text of one item